requirements.txt
gspread
requests
pandas
schedule
cryptography>=41.0.0
python-binance>=1.0.15
oauth2client